+++ example/integration_test/_integration_testing.dart
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'package:example/src/controller.dart' show Controller;
+import 'package:example/src/controller.dart';
 
 import 'package:example/src/view.dart';
 
@@ -14,6 +14,20 @@ const _location = '========================== test_example_app.dart';
 
 ///
 Future<void> integrationTesting(WidgetTester tester) async {
+  // A Singleton pattern allows for unit testing.
+  final Controller con = Controller();
+
+  // You can retrieve a State object the Controller has collected so far.
+  StateX state = con.stateOf<Page1>()! as StateX;
+
+  /// State object's first build or will be its first build
+  expect(con.statex!.firstBuild, isTrue, reason: _location);
+
+  // Remove the Debug Banner
+  DevTools().debugShowCheckedModeBanner = false;
+  con.appStateX?.setState(() {});
+  await tester.pumpAndSettle(const Duration(seconds: 1));
+
   //
   await testPage1(tester);
 
@@ -21,17 +35,12 @@ Future<void> integrationTesting(WidgetTester tester) async {
   await tester.tap(find.byKey(const Key('Page 2')));
   await tester.pumpAndSettle(const Duration(milliseconds: 200));
 
-  StateX state;
-
-  // A Singleton pattern allows for unit testing.
-  final Controller? con = Controller();
-
   // Test the controllers move across different State objects.
-  state = con?.statex!.firstState as StateX;
+  state = con.statex!.firstState as StateX;
 
   expect(state, isA<AppStateX>(), reason: _location);
 
-  state = con?.statex!.lastState as StateX;
+  state = con.statex!.lastState as StateX;
 
   expect(state, isA<Page2State>(), reason: _location);
 
@@ -47,6 +56,10 @@ Future<void> integrationTesting(WidgetTester tester) async {
     await tester.tap(find.byKey(const Key('+')));
     await tester.pumpAndSettle();
   }
+
+  /// A flag. Noting if the function setBuilder() used.
+  /// Widget setBuilder(MaybeBuildWidgetType? builder)
+  expect(con.setBuilderUsed, isTrue, reason: _location);
 
   expect(find.text((count).toString()), findsOneWidget, reason: _location);